perf: move Bun.build() calls into subprocesses to avoid EISDIR bug
Bun has a bug where multiple Bun.build() calls in the same process
that touch overlapping transitive deps (devalue, svelte, mochi-framework)
produce spurious EISDIR errors. Move the client bundle build and inline
web component builds into separate subprocesses so each Bun.build() runs
in its own process with fresh bundler state.

- Extract shared helpers (applyUserPreprocessors, createMarkdownLoader,
  formatBuildMessages) into svelteCompileHelpers.ts
- Add buildClientWorker.ts subprocess that reconstructs the client plugin
  and runs Bun.build() in isolation
- Add buildInlineWorker.ts for inline web component builds
- Replace inline markdown option with markdownConfigPath (file-based
  config that both main process and subprocesses can load)
- Add loadMarkdownConfig.ts shared helper
